+++ src/mcp_edge/transports/__init__.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from .base import Transport, TransportError
+from .ble import BLETransport
 from .loopback import FrameHandler, LoopbackTransport
 from .serial import SerialTransport
 
@@ -12,4 +13,5 @@ __all__ = [
     "LoopbackTransport",
     "FrameHandler",
     "SerialTransport",
+    "BLETransport",
 ]